backend/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
requests
pyrebase4
python-dotenv
